Error log updated for: (#80)
* Error log updated for:
"If the xcode archive is success, we have the archive required codesigning files. Then codesigndoc asks for the distribution method, if method selected, which codesigning files are not exist on the developer's machine it mentions the fail, and asks if do you want to export other files? If no selected it finishes with strange error: no ipa export code sign groups collected"

* PR clean - fix (+1 squashed commit)
Squashed commits:
[b19a1ac] PR clean - fix